Moved more test packages modified location and it's realtions to other 2 entities
