+++ projDelivera/backend/delivera/src/main/java/ua/tqs/delivera/models/Location.java
@@ -1,11 +1,6 @@
 package ua.tqs.delivera.models;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 import ua.tqs.delivera.datamodels.LocationDTO;
 
@@ -20,7 +15,8 @@ public class Location {
     private double latitude;
     @Column
     private double longitude;
-
+    
+    
     public Location(){}
     
     public Location(double latitude, double longitude){
