+++ projDelivera/backend/delivera/src/main/java/ua/tqs/delivera/services/RiderService.java
@@ -4,8 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
+import ua.tqs.delivera.models.Order;
 import ua.tqs.delivera.models.Rider;
 import ua.tqs.delivera.repositories.RiderRepository;
+
+import java.util.List;
 
 @Service
 public class RiderService {
@@ -17,5 +20,15 @@ public class RiderService {
         if (!riderRepo.findByEmail(rider.getEmail()).isPresent())
             return riderRepo.save(rider);
         throw new DuplicateKeyException("Email already in use");
+    }
+    
+    
+    /*
+    * Get All orders for rider with id = riderID
+    * return a List of orders
+    *
+    */
+    public List<Order> getAllOrdersForRider( Long riderID){
+        return null;
     }
 }
