+++ projDelivera/backend/delivera/src/test/java/ua/tqs/delivera/controllertests/DeliveraControllerTests.java
@@ -1,4 +1,4 @@
-package ua.tqs.delivera.ControllerTests;
+package ua.tqs.delivera.controllertests;
 
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
