+++ projDelivera/backend/delivera/src/test/java/ua/tqs/delivera/repositoriestests/RiderRepositoryTest.java
@@ -1,4 +1,4 @@
-package ua.tqs.delivera.RepositoryTests;
+package ua.tqs.delivera.repositoriestests;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Order;
@@ -9,26 +9,28 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
+
+import ua.tqs.delivera.models.Location;
+import ua.tqs.delivera.models.Rider;
+import ua.tqs.delivera.repositories.RiderRepository;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-import ua.tqs.delivera.models.Location;
-import ua.tqs.delivera.repositories.LocationRepository;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
 @DataJpaTest
 @Testcontainers
-//To use the same datasource as the regular application
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class LocationRepositoryTest {
+class RiderRepositoryTest {
     @Autowired
     private TestEntityManager entityManager;
 
     @Autowired
-    private LocationRepository locationRepo;
+    private RiderRepository riderRepo;
 
+    private Rider rider;
     private Location location;
 
     @Container
@@ -47,28 +49,31 @@ public class LocationRepositoryTest {
     @BeforeEach
     void setup(){
         location = new Location(40.85, 25.9999);
+        rider = new Rider("ma@gmail.com", "Manuel Antunes", "migant", true, location, 0, 0);
     }
 
     @Order(1)
     @Test
-    void whenSaveValidLocation_thenReturnLocation(){
-        Location location_saved = entityManager.persistAndFlush(location);
-        assertThat(location_saved).isEqualTo(location);
+    void whenSaveRider_thenReturnRider(){
+        entityManager.persistAndFlush(location);
+        Rider rider_saved = entityManager.persistAndFlush(rider);
+        assertThat(rider_saved).isEqualTo(rider);
     }
 
     @Order(2)
     @Test
     void whenFindRiderByExistingId_thenReturnRider(){
         entityManager.persistAndFlush(location);
+        entityManager.persistAndFlush(rider);
 
-        Location foundLocation = locationRepo.findById(location.getId()).get();
-        assertThat(foundLocation).isEqualTo(location);
+        Rider foundRider = riderRepo.findById(rider.getRiderId()).get();
+        assertThat(foundRider).isEqualTo(rider);
     }
 
     @Order(3)
     @Test
     void whenInvalidId_thenReturnNull(){
-        Location location_saved = locationRepo.findById(-1L).orElse(null);
-        assertThat(location_saved).isNull();
+        Rider rider_saved = riderRepo.findById(-1L).orElse(null);
+        assertThat(rider_saved).isNull();
     }
 }
