+++ projDelivera/backend/delivera/src/test/java/ua/tqs/delivera/servicestests/AdminServiceTest.java
@@ -1,4 +1,4 @@
-package ua.tqs.delivera.ServiceTests;
+package ua.tqs.delivera.servicestests;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
